fix: move heroEmoji to lib so server components can call it
heroEmoji was defined in StoreCard (a "use client" file). Calling it
from the store profile page (a server component) threw:
"cannot invoke a client function from the server".

Move the function and CATEGORY_EMOJI map to lib/store-hero-images.ts
(no "use client" directive). StoreCard re-exports it from there so
existing imports keep working.
